+++ sections/shorting/ShortingCard/components/CRatioSelector/CRatioSelector.tsx
@@ -20,15 +20,7 @@ import {
 import CaretDownIcon from 'assets/svg/app/caret-down.svg';
 import useCollateralShortContractInfoQuery from 'queries/collateral/useCollateralShortContractInfoQuery';
 
-export type ShortCRatioLevel = 'safe' | 'safeMax' | 'highRisk';
-
-export const SHORT_C_RATIO: Record<ShortCRatioLevel, number> = {
-	safe: 2,
-	safeMax: 1.65,
-	highRisk: 1.55,
-};
-
-export const shortCRatios = Object.entries(SHORT_C_RATIO) as [ShortCRatioLevel, number][];
+import { shortCRatios } from './constants';
 
 type CRatioSelectorProps = {};
